+++ deprecated/examples/vue-quickstart/src/pages/[...all].vue
@@ -1,5 +1,0 @@
-<template>
-  <div>
-    Not Found
-  </div>
-</template>