+++ deprecated/packages/hasura-auth-js/nhost/migrations/default/1644527455723_create_table_public_test/up.sql
@@ -1,2 +1,0 @@
-CREATE TABLE "public"."test" ("id" uuid NOT NULL DEFAULT gen_random_uuid(), PRIMARY KEY ("id") );
-CREATE EXTENSION IF NOT EXISTS pgcrypto;