+++ flake.nix
@@ -57,6 +57,10 @@
           inherit self pkgs nix-filter nixops-lib;
         };
 
+        stripe-graphql-jsf = import ./packages/stripe-graphql-js/project.nix {
+          inherit self pkgs nix-filter nixops-lib;
+        };
+
         nixopsf = import ./nixops/project.nix {
           inherit self pkgs nix2containerPkgs nix-filter nixops-lib;
         };
@@ -84,6 +88,7 @@
           guides = guidesf.check;
           docs = docsf.check;
           nhost-js = nhost-jsf.check;
+          stripe-graphql-js = stripe-graphql-jsf.check;
           nixops = nixopsf.check;
           postgres = postgresf.check;
           storage = storagef.check;
@@ -122,6 +127,14 @@
               golangci-lint
               gqlgenc
               oapi-codegen
+              mockgen
+              sqlc
+              vacuum-go
+
+              # others
+              postgresql_18-client
+              bun
+
 
               # docs
               vale
@@ -171,6 +184,7 @@
           guides = guidesf.devShell;
           docs = docsf.devShell;
           nhost-js = nhost-jsf.devShell;
+          stripe-graphql-js = stripe-graphql-jsf.devShell;
           nixops = nixopsf.devShell;
           postgres = postgresf.devShell;
           storage = storagef.devShell;
@@ -189,6 +203,7 @@
           demos = demosf.package;
           guides = guidesf.package;
           nhost-js = nhost-jsf.package;
+          stripe-graphql-js = stripe-graphql-jsf.package;
           nixops = nixopsf.package;
           nixops-docker-image = nixopsf.dockerImage;
           postgres-pg16 = postgresf.packages.pg16-package;